Updated few modifications
